services/api-gateway/pyproject.toml
[project]
name = "api-gateway"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.115.13",
    "uvicorn==0.34.0",
    "httpx==0.28.1",
    "opentelemetry-distro>=0.50b0",
    "opentelemetry-exporter-otlp>=1.30",
    "opentelemetry-instrumentation-fastapi>=0.50b0",
    "opentelemetry-instrumentation-httpx>=0.50b0",
    "opentelemetry-instrumentation-logging>=0.50b0",
    "opentelemetry-instrumentation-system-metrics>=0.50b0",
    "pyroscope-io>=0.8.7",
]
